backend/auth/requirements.txt
psycopg2-binary==2.9.9
passlib==1.7.4
bcrypt==4.0.1